address use of numpy/cupy
